requirements.txt
firecrawl>=4.20.0
python-dotenv>=1.0.0
streamlit>=1.40.0
pandas>=2.2.0
requests>=2.31.0
anthropic>=0.40.0
